+++ CHANGELOG.md
@@ -18,8 +18,8 @@ Improvements
 
 - Remove sensitive information.
 - Improve log.
-- Add `ChatGroupEventListener.onMemberAttributesChanged` notification.
-- Adds the `fetchHistoryMessages` chat manager api.
+- Adds `ChatGroupEventListener.onMemberAttributesChanged` notification.
+- Updates the `fetchHistoryMessages` chat manager api.
 
 Issues fixed